requirements.txt
boto3==1.34.34
requests==2.31.0
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1